requirements.txt
asgiref
black
boto3
botocore
cfgv
click
coverage
distlib
Django
django-cors-headers
django-storages
djangorestframework
exceptiongroup
filelock
flake8
flake8-django
identify
iniconfig
isort
jmespath
mccabe
mypy-extensions
nodeenv
packaging
pathspec
platformdirs
pluggy
pre-commit
pycodestyle
pyflakes
pytest
pytest-cov
pytest-django
python-dateutil
python-dotenv
pytz
PyYAML
s3transfer
six
sqlparse
tomli
urllib3
virtualenv
drf-yasg
redis
celery
django-celery-beat
django-celery-results
requests
psycopg[binary,pool]
psycopg2-binary
pyjwt
oauthlib